+++ rs/src/tags.rs
@@ -364,6 +364,12 @@ pub struct PlaceObject {
   pub clip_actions: Option<Vec<ClipAction>>,
 }
 
+#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub struct Protect {
+  pub password: String,
+}
+
 #[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
 pub struct RemoveObject {