+++ ts/src/lib/tags/index.ts
@@ -26,6 +26,7 @@ export { $FrameLabel, FrameLabel } from "./frame-label";
 export { $ImportAssets, ImportAssets } from "./import-assets";
 export { $Metadata, Metadata } from "./metadata";
 export { $PlaceObject, PlaceObject } from "./place-object";
+export { $Protect, Protect } from "./protect";
 export { $RemoveObject, RemoveObject } from "./remove-object";
 export { $ScriptLimits, ScriptLimits } from "./script-limits";
 export { $SetBackgroundColor, SetBackgroundColor } from "./set-background-color";